teste_3/requirements.txt
greenlet==2.0.2
mysql-connector-python==8.1.0
netifaces==0.11.0
pika==1.3.2
protobuf==4.21.12
psutil==5.9.5
python-decouple==3.8
SQLAlchemy==2.0.21
typing-extensions==4.8.0
